+++ src/main/java/me/villagerunknown/immortalvillagers/feature/resetVillagerTradesFeature.java
@@ -5,10 +5,13 @@ import me.villagerunknown.platform.util.EntityUtil;
 import net.fabricmc.fabric.api.event.player.UseEntityCallback;
 import net.minecraft.entity.EntityType;
 import net.minecraft.entity.passive.VillagerEntity;
+import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.Items;
 import net.minecraft.particle.ParticleEffect;
 import net.minecraft.particle.ParticleTypes;
+import net.minecraft.sound.SoundCategory;
+import net.minecraft.sound.SoundEvents;
 import net.minecraft.text.Text;
 import net.minecraft.util.ActionResult;
 import net.minecraft.village.TradeOfferList;
@@ -20,6 +23,11 @@ import java.util.List;
 public class resetVillagerTradesFeature {
 	
 	public static String RESET_STRING = Immortalvillagers.CONFIG.villagerTradesResetItemName;
+	
+	public static List<Item> RESET_ITEMS = List.of(
+			Items.EMERALD,
+			Items.EMERALD_BLOCK
+	);
 	
 	public static List<VillagerProfession> NO_RESET_PROFESSIONS = List.of(
 			VillagerProfession.NONE,
@@ -41,7 +49,7 @@ public class resetVillagerTradesFeature {
 				VillagerEntity villager = (VillagerEntity) entity;
 				VillagerProfession profession = villager.getVillagerData().getProfession();
 				
-				if( !NO_RESET_PROFESSIONS.contains( profession ) && itemStack.getItem().equals( Items.PAPER ) && itemStack.getName().getString().equalsIgnoreCase( RESET_STRING ) ) {
+				if( !NO_RESET_PROFESSIONS.contains( profession ) && RESET_ITEMS.contains( itemStack.getItem() ) && itemStack.getName().getString().equalsIgnoreCase( RESET_STRING ) ) {
 					itemStack.decrementUnlessCreative( 1, player );
 					
 					villager.setOffers(null);
@@ -58,8 +66,8 @@ public class resetVillagerTradesFeature {
 					
 					villager.setOffers( offers );
 					
-					villager.playWorkSound();
-					EntityUtil.spawnParticles( villager, 1.5F, ParticleTypes.HAPPY_VILLAGER, 10, 0.05, 0.05, 0.05, 0.5);
+					world.playSoundFromEntity( villager, SoundEvents.ENTITY_VILLAGER_TRADE, SoundCategory.NEUTRAL, 1, 1 );
+					EntityUtil.spawnParticles( villager, 1.5F, ParticleTypes.HAPPY_VILLAGER, 10, 0.5, 0.5, 0.5, 0.5);
 					
 					return ActionResult.SUCCESS;
 				} // if